updating docs
