+++ packages/signals/pubspec.yaml
@@ -2,7 +2,7 @@ name: signals
 description: "signal, computed and effect built with dart and extensions for flutter"
 repository: https://github.com/rodydavis/signals.dart
 homepage: https://rodydavis.github.io/signals.dart/
-version: 1.2.1
+version: 1.2.2
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
