+++ website/src/content/docs/reference/watch.md
@@ -3,25 +3,57 @@ title: Watch
 description: Watch a signal for changes in Flutter
 ---
 
-To watch a signal for changes in Flutter, use the `watch` extension.
+## Watch
+
+To watch a signal for changes in Flutter, use the `Watch` widget. This will only rebuild this widget method and not the entire widget tree.
 
 ```dart
 final signal = signal(10);
 ...
 @override
 Widget build(BuildContext context) {
-  final count = signal.watch(context);
-  return Text('$count');
+  return Watch(() => Text('$signal'));
 }
 ```
 
-## Rebuilds
+## SignalWidget
+
+If using a `StatelessWidget` and you want to watch a signal for changes rebuilding the entire widget, use the `SignalWidget` widget.
+
+```diff
+final signal = signal(10);
+...
+- class MyWidget extends StatelessWidget {
++ class MyWidget extends SignalWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Text('$signal');
+  }
+}
+```
+
+## .watch(context)
+
+If you need to map to a widget property use the `watch` extension method. This will infer the type and subscribe to the signal.
+
+```dart
+final fontSize = signal(10);
+...
+@override
+Widget build(BuildContext context) {
+  return Text('Hello World',
+    style: TextStyle(fontSize:fontSize.watch(context)),
+  );
+}
+```
+
+### Rebuilds
 
 To protect against unnecessary rebuilds, the `watch` extension will only subscribe once to the nearest element and mark the widget as dirty.
 
 This means that if you have multiple widgets that are watching the same signal, only the first one will be subscribed to the signal and multiple updates will be batched together.
 
-It is also possible to isolate the rebuilds with the `Builder` widget.
+It is also possible to isolate the rebuilds with the `Builder` widget, however it is recommended to use `Watch` or `SignalWidget` instead.
 
 ```dart
 final signal = signal(10);
@@ -54,7 +86,6 @@ final computed = computed(() => signal.value.a);
 ...
 @override
 Widget build(BuildContext context) {
-  final aCount = computed.watch(context);
-  return Text('$aCount');
+  return Watch(() => Text('$computed'));
 }
 ```
